cleanup tests directory
