+++ tests/plugins/cassettes/test_calculate_checkout_line_total[True-15.00-18.45-False].yaml
