+++ tests/plugins/cassettes/test_calculate_checkout_subtotal[False-40.65-50.00-True].yaml
