+++ tests/plugins/cassettes/test_calculate_checkout_subtotal[True-20.33-25.00-True].yaml
