+++ tests/plugins/cassettes/test_calculate_checkout_subtotal[True-25.00-30.75-False].yaml
