+++ tests/plugins/cassettes/test_calculate_checkout_total[False-29.52-37.00-3.0-True].yaml
