+++ tests/plugins/cassettes/test_calculate_checkout_total[False-37.00-37.00-3.0-True].yaml
