+++ tests/plugins/cassettes/test_calculate_checkout_total[False-40.00-40.00-0.0-False].yaml
