+++ tests/plugins/cassettes/test_calculate_checkout_total[True-20.00-20.00-5.0-False].yaml
