+++ tests/plugins/cassettes/test_calculate_checkout_total[True-20.00-25.75-5.0-False].yaml
