+++ tests/plugins/cassettes/test_plugin_uses_configuration_from_db.yaml
